Fix Link component references
Change-type: patch
