+++ lib/Link/Link.jsx
@@ -10,43 +10,6 @@ import React from 'react'
 import {
 	Link as RenditionLink
 } from 'rendition'
-import {
-	isRelativeUrl,
-	isLocalUrl,
-	toRelativeUrl
-} from '../services/helpers'
-
-export const getLinkProps = (href) => {
-	if (isRelativeUrl(href)) {
-		return {
-			append: href.replace(/^\//, '')
-		}
-	}
-	return (isLocalUrl(href))
-		? {
-			append: toRelativeUrl(href).replace(/^\//, '')
-		}
-		: {
-			to: href,
-			blank: true
-		}
-}
-
-export const linkComponentOverride = ({
-	blacklist
-}) => {
-	return ({
-		href, ...rest
-	}) => {
-		if (_.some(blacklist, (url) => {
-			return href.match(url)
-		})) {
-			return null
-		}
-		const linkProps = getLinkProps(href)
-		return <Link {...rest} {...linkProps} />
-	}
-}
 
 export class Link extends React.Component {
 	constructor (props) {
