+++ lib/Link/index.jsx
@@ -4,16 +4,50 @@
  * Proprietary and confidential.
  */
 
+import React from 'react'
+import _ from 'lodash'
 import {
 	withRouter
 } from 'react-router-dom'
 import {
 	Link as InnerLink
 } from './Link'
-
-export {
-	getLinkProps,
-	linkComponentOverride
-} from './Link'
+import {
+	isRelativeUrl,
+	isLocalUrl,
+	toRelativeUrl
+} from '../services/helpers'
 
 export const Link = withRouter(InnerLink)
+
+export const getLinkProps = (href) => {
+	if (isRelativeUrl(href)) {
+		return {
+			append: href.replace(/^\//, '')
+		}
+	}
+	return (isLocalUrl(href))
+		? {
+			append: toRelativeUrl(href).replace(/^\//, '')
+		}
+		: {
+			to: href,
+			blank: true
+		}
+}
+
+export const linkComponentOverride = ({
+	blacklist
+}) => {
+	return ({
+		href, ...rest
+	}) => {
+		if (_.some(blacklist, (url) => {
+			return href.match(url)
+		})) {
+			return null
+		}
+		const linkProps = getLinkProps(href)
+		return <Link {...rest} {...linkProps} />
+	}
+}
